+++ src/pathfinders.h
@@ -64,7 +64,7 @@ class PathFinder {
         void init_arrays (
                 std::vector<double>& d,
                 std::vector<double>& w,
-                std::vector<int>& prev,
+                std::vector<long int>& prev,
                 bool *m_open_vec,
                 bool *m_closed_vec,
                 const size_t v,
@@ -73,7 +73,7 @@ class PathFinder {
                 const DGraphEdge *edge,
                 std::vector<double>& d,
                 std::vector<double>& w,
-                std::vector<int>& prev,
+                std::vector<long int>& prev,
                 bool *m_open_vec,
                 const bool *m_closed_vec,
                 const size_t &v0);
@@ -81,7 +81,7 @@ class PathFinder {
                 const DGraphEdge *edge,
                 std::vector<double>& d,
                 std::vector<double>& w,
-                std::vector<int>& prev,
+                std::vector<long int>& prev,
                 bool *m_open_vec,
                 const bool *m_closed_vec,
                 const size_t &v0,
@@ -90,24 +90,24 @@ class PathFinder {
         void Dijkstra (
                 std::vector<double>& d,
                 std::vector<double>& w,
-                std::vector<int>& prev,
+                std::vector<long int>& prev,
                 const size_t v0,
                 const std::vector <size_t> &to_index);
         void DijkstraLimit (
                 std::vector<double>& d,
                 std::vector<double>& w,
-                std::vector<int>& prev,
+                std::vector<long int>& prev,
                 const size_t v0,
                 const double &dlim);
         void AStar (std::vector<double>& d,
                 std::vector<double>& w,
-                std::vector<int>& prev,
+                std::vector<long int>& prev,
                 const std::vector<double>& heur,
                 const size_t v0,
                 const std::vector <size_t> &to_index);
         void Dijkstra_set (std::vector <double>& d,
                 std::vector<double>& w,
-                std::vector<int>& prev,
+                std::vector<long int>& prev,
                 size_t v0);
         void Centrality_vertex (
                 std::vector <double>& cent,
